+++ src/axom/bump/PrimalAdaptor.hpp
@@ -310,14 +310,15 @@ struct AdaptPolyhedron<TopologyView, CoordsetView, true>
  *
  * \tparam TopologyView The topology view type.
  * \tparam CoordsetView The coordset view type.
+ * \tparam MAX_VERTS_2D The maximum number of vertices allowed in a polygon.
  * \tparam makeFaces Whether to make faces for polyhedral shapes or to make primal::Polyhedron.
  */
-template <typename TopologyView, typename CoordsetView, bool makeFaces = false>
+template <typename TopologyView, typename CoordsetView, int MAX_VERTS_2D = 12, bool makeFaces = false>
 struct PrimalAdaptor
 {
   using value_type = typename CoordsetView::value_type;
   using Polygon =
-    axom::primal::Polygon<value_type, CoordsetView::dimension(), axom::primal::PolygonArray::Static>;
+    axom::primal::Polygon<value_type, CoordsetView::dimension(), axom::primal::PolygonArray::Static, MAX_VERTS_2D>;
   using Tetrahedron = axom::primal::Tetrahedron<value_type, CoordsetView::dimension()>;
   using Hexahedron = axom::primal::Hexahedron<value_type, CoordsetView::dimension()>;
   using Polyhedron =