+++ src/axom/mir/EquiZAlgorithm.hpp
@@ -50,6 +50,11 @@ namespace mir
  * \tparam TopologyView A topology view to be used for accessing zones in the mesh.
  * \tparam CoordsetView A coordset view that accesses coordinates as primal::Point.
  * \tparam MatsetView A matset view that interfaces to the Blueprint material set.
+ *
+ * \note This algorithm typically produces unstructured meshes of zoo elements.
+ *       However, if the input matset contains only "clean" zones consisting of 1
+ *       material per zone then the input coordset, topology, and matset will be
+ *       copied to the output. In that case, the types will depend on the input types.
  */
 template <typename ExecSpace, typename TopologyView, typename CoordsetView, typename MatsetView>
 class EquiZAlgorithm : public axom::mir::MIRAlgorithm
@@ -71,6 +76,7 @@ public:
     , m_topologyView(topoView)
     , m_coordsetView(coordsetView)
     , m_matsetView(matsetView)
+    , m_selectionKey("selectedZones")
   { }
 
   /// Destructor
@@ -130,7 +136,6 @@ protected:
     // Come up with lists of clean/mixed zones.
     axom::Array<axom::IndexType> cleanZones, mixedZones;
     makeZoneLists(n_options_copy, cleanZones, mixedZones);
-    SLIC_ASSERT((cleanZones.size() + mixedZones.size()) == m_topologyView.numberOfZones());
     SLIC_INFO(
       axom::fmt::format("cleanZones: {}, mixedZones: {}", cleanZones.size(), mixedZones.size()));
 
@@ -167,7 +172,7 @@ protected:
       }
 
       // Process the mixed part of the mesh. We select just the mixed zones.
-      n_options_copy["selectedZones"].set_external(mixedZones.data(), mixedZones.size());
+      n_options_copy[m_selectionKey].set_external(mixedZones.data(), mixedZones.size());
       n_options_copy["newNodesField"] = newNodesFieldName();
       processMixedZones(n_root_topo,
                         n_root_coordset,
@@ -225,21 +230,46 @@ protected:
     }
     else if(cleanZones.size() > 0 && mixedZones.size() == 0)
     {
-      // There were no mixed zones. We can copy the input to the output.
+      // There were no mixed zones.
+
+      if(!n_options_copy.has_path(m_selectionKey))
       {
+        // We can copy the input to the output (no selected zones).
         AXOM_ANNOTATE_SCOPE("copy");
         utils::copy<ExecSpace>(n_newCoordset, n_coordset);
         utils::copy<ExecSpace>(n_newTopo, n_topo);
         utils::copy<ExecSpace>(n_newFields, n_fields);
         utils::copy<ExecSpace>(n_newMatset, n_matset);
-      }
 
-      // Add an originalElements array.
-      const std::string originalElementsField(axom::bump::Options(n_options).originalElementsField());
-      addOriginal(n_newFields[originalElementsField],
-                  n_topo.name(),
-                  "element",
-                  m_topologyView.numberOfZones());
+        // Add an originalElements array.
+        const std::string originalElementsField(
+          axom::bump::Options(n_options).originalElementsField());
+        addOriginal(n_newFields[originalElementsField],
+                    n_newTopo.name(),
+                    "element",
+                    m_topologyView.numberOfZones());
+      }
+      else
+      {
+        // Make the clean mesh of only the selected zones
+
+        conduit::Node n_root;
+        n_root[localPath(n_coordset)].set_external(n_coordset);
+        n_root[localPath(n_topo)].set_external(n_topo);
+        n_root[localPath(n_matset)].set_external(n_matset);
+
+        conduit::Node n_cleanOutput;
+        makeCleanZones(n_root, n_topo.name(), n_options_copy, cleanZones.view(), n_cleanOutput);
+
+        // Move n_cleanOutput objects into the supplied nodes.
+        n_newCoordset.move(n_cleanOutput[localPath(n_newCoordset)]);
+        n_newTopo.move(n_cleanOutput[localPath(n_newTopo)]);
+        n_newMatset.move(n_cleanOutput[localPath(n_newMatset)]);
+        if(n_cleanOutput.has_path("fields"))
+        {
+          n_newFields.move(n_cleanOutput["fields"]);
+        }
+      }
     }
 #else
     // Handle all zones via MIR.
@@ -272,17 +302,22 @@ protected:
     namespace utils = axom::bump::utilities;
     axom::bump::ZoneListBuilder<ExecSpace, TopologyView, MatsetView> zlb(m_topologyView,
                                                                          m_matsetView);
-    if(n_options.has_child("selectedZones"))
+    [[maybe_unused]] axom::IndexType expectedSize = 0;
+    if(n_options.has_child(m_selectionKey))
     {
       auto selectedZonesView =
-        utils::make_array_view<axom::IndexType>(n_options.fetch_existing("selectedZones"));
+        utils::make_array_view<axom::IndexType>(n_options.fetch_existing(m_selectionKey));
       zlb.execute(m_coordsetView.numberOfNodes(), selectedZonesView, cleanZones, mixedZones);
+      expectedSize = selectedZonesView.size();
     }
     else
     {
       zlb.execute(m_coordsetView.numberOfNodes(), cleanZones, mixedZones);
+      expectedSize = m_topologyView.numberOfZones();
     }
     // _bump_utilities_zlb_end
+
+    SLIC_ASSERT((cleanZones.size() + mixedZones.size()) == expectedSize);
   }
 
   /*!
@@ -568,9 +603,9 @@ protected:
         // In later iterations, we do not want to pass selectedZones through
         // since they are only valid on the current input topology. Also, if they
         // were passed then the new topology only has those selected zones.
-        if(n_options.has_child("selectedZones"))
+        if(n_options.has_child(m_selectionKey))
         {
-          n_options.remove("selectedZones");
+          n_options.remove(m_selectionKey);
         }
       }
       else
@@ -995,10 +1030,10 @@ protected:
     options["inside"] = 1;
     options["outside"] = 1;
     options["colorField"] = colorField;
-    if(n_options.has_child("selectedZones"))
+    if(n_options.has_child(m_selectionKey))
     {
       // Pass selectedZones along in the clip options, if present.
-      options["selectedZones"].set_external(n_options.fetch_existing("selectedZones"));
+      options[m_selectionKey].set_external(n_options.fetch_existing(m_selectionKey));
     }
     if(n_options.has_child("fields"))
     {
@@ -1150,6 +1185,7 @@ private:
   TopologyView m_topologyView;
   CoordsetView m_coordsetView;
   MatsetView m_matsetView;
+  std::string m_selectionKey;
 };
 
 }  // end namespace mir